Transliteration
Removal of line breaks in foreign tags
	s/(?<=lang="ar">)([^<]*)\n\t+([^<]*)(?=<\/)/$1$2/g

Correction of shadda error where a 3 is shown as third radical instead of duplicating the second radical:
	s/(lang="ar">[^<]*)(.)(.)3([^<]*<\/)/$1$2$3$2$4/g

Then completely transliterated the file.

There is still errors in orth tags which have '^' at the wrong place, so it was joined with 'A'. (in 131 places). Part of it will be corrected with the aor. expansion the others all seem to be orth tags of verbforms which were incorrectly created.